Save shipping method in Cart during checkout
